fix(deps): remove openssl dep and downgrade lzma
reqwest, which is where the openssl requirement came from, now uses the rustls-tls feature which removes the openssl requirement.

downgrading liblzma to 0.2, which uses lzma 5.4. The previous liblzma, 0.3, uses lzma 5.6, which is not availble easily on most OSs due to major bug, so the downgrade makes it easier to compile nohuman for other architectures.